[#1] code cleanup & add player control logic
merge 옵션, 등등 추가함
