+++ Assets/Scenes/NetcodeTest.unity
@@ -5977,6 +5977,22 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: -278549503891398555, guid: 2428677ba9b5e8f4f80d334268ec0774, type: 3}
+      propertyPath: Interpolate
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: -278549503891398555, guid: 2428677ba9b5e8f4f80d334268ec0774, type: 3}
+      propertyPath: PositionThreshold
+      value: 1e-10
+      objectReference: {fileID: 0}
+    - target: {fileID: -278549503891398555, guid: 2428677ba9b5e8f4f80d334268ec0774, type: 3}
+      propertyPath: UseUnreliableDeltas
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: -278549503891398555, guid: 2428677ba9b5e8f4f80d334268ec0774, type: 3}
+      propertyPath: PositionMaxInterpolationTime
+      value: 0.01
+      objectReference: {fileID: 0}
     - target: {fileID: 3226371800214689686, guid: 2428677ba9b5e8f4f80d334268ec0774, type: 3}
       propertyPath: GlobalObjectIdHash
       value: 2711001984
