+++ Assets/Scripts/Network/PlayerConstants.cs
@@ -1,0 +1,13 @@
+﻿namespace Network {
+    public enum PlayerAction {
+        RightMove,
+        LeftMove,
+        UpMove,
+        DownMove,
+        Jump
+    }
+
+    public enum MergeStrategy {
+        Latest
+    }
+}
